graceful shutdown

+++ install/docker-stop.sh
@@ -1,6 +1,3 @@
 #!/bin/bash
 
-/opt/eva/sbin/eva-control stop
-killall tail
-kill `cat /var/run/start.pid`
-rm -f /var/run/start.pid
+[ -f /var/run/start.pid ] && kill `cat /var/run/start.pid` || killall bash
